Fix minor code issue pointed out during talk at LCA
This commit fixes an issue pointed out during the questions when I gave
the talk at LCA. [1] They correctly pointed out that I was passing the
classical register to the oracle circuit definition even when it wasn't
used. During the talk I missed what he was mentioning and didn't
understand the question. This commit fixes that to avoid confusion so
the oracle circuit definition only relies on the registers it uses.
While passing an extra register to the definition doesn't do anything if
it's not used in the circuit, this was a source of confusion so this
addresses that.

[1] https://youtu.be/th4TDYX6xoc?t=2402